cleaned up build, seperated sass task
